Simplified part two of day 2 in Haskell
I think this should be a little bit more efficienct since it's more
lazy:

* instead of counting the number of differences between two IDs, we
(lazily) build a list of differences and then don't bother checking
further if we find two or more differences.

* instead of counting the number of differences for all pairs, and theen
selecting the minimum, we assert that there is one pair with exactly one
difference and stop searching as soon as we find it.