Day 3 in Rust
This is not as efficient as it could: the program recomputes the set of
all covered cells and the number of times each cell is covered
unnecessarily. However, a release build needs just about 130ms on my
laptop, so this is good enough for me. I feel that more sharing would
make the part_one/part_two signatures uglier.